+++ src/test/components/static/staticResult.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import { CLASS_NAME, VALUE } from './export';
 
+const CONSTANT_IN_FILE = 1;
+
 function Component() {
   const baseClass: string = 'my-class-2';
   return <Extracted baseClass={baseClass} />;
@@ -21,6 +23,8 @@ function Extracted({ baseClass }: ExtractedProps) {
       <input min={0} max={VALUE} className={CLASS_NAME + 'text-sm'} />
       <button className={CLASS_NAME + baseClass} />
       <div className='another-class'>Test2</div>
+      <div>{CONSTANT_IN_FILE}</div>
     </div>
   );
 }
+